Exceptions cunstom

+++ src/main/java/com/githunb/ailton78/crudeDeClientes/services/ClientService.java
@@ -3,11 +3,17 @@ package com.githunb.ailton78.crudeDeClientes.services;
 import com.githunb.ailton78.crudeDeClientes.dto.ClientDTO;
 import com.githunb.ailton78.crudeDeClientes.entities.Client;
 import com.githunb.ailton78.crudeDeClientes.repositories.ClientRepository;
+import com.githunb.ailton78.crudeDeClientes.services.exceptions.DatabaseException;
+import com.githunb.ailton78.crudeDeClientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 
 @Service
 public class ClientService {
@@ -16,13 +22,14 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado!"));
         return  new ClientDTO(client);
     }
     @Transactional(readOnly = true)
-    public List<ClientDTO> findAll(){
-        List<Client> list = repository.findAll();
-        return list.stream().map(ClientDTO::new).toList();
+    public Page<ClientDTO> findAll(Pageable pageable){
+        Page<Client> list = repository.findAll(pageable);
+        return list.map(ClientDTO::new);
     }
     @Transactional
     public ClientDTO insert(ClientDTO dto){
@@ -33,14 +40,27 @@ public class ClientService {
     }
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        repository.save(entity);
-        return new ClientDTO(entity);
+        try{
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            repository.save(entity);
+            return new ClientDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
